pyproject.toml
[project]
name = "rag-support-assistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.2",
    "datasets>=4.6.1",
    "fastapi>=0.135.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
